completely revamp how we load maps
Instead of a single .tfimap file and a state machine to deal with different inner formats
we define a manifest with key metadata and a set of csv files, one per each layer

This simplifies map import from Tiled and support for multiple layers.

The client is the same so we still only send layer 0.