stair logic fixes, bump to v 20.12.27a
Visual: we now boost the z order in more cases while in the stairs, that seems to solve the last clipping problems.

Debu message: now we test the right condtion so we don't get duplicated messages.

Map wise we are renaming the levels so they match the z_level, with T0 being z_order zero and TG being -1.

There are a couple of client fixes needed which will bump to same version.