Player state refactor
Significant refactor of the player state, which is now a separate readonly struct which the GameLogic cannot mutate directly but can construct a mutated version of it.

Logic is cleaner and opens the way to unify the places we send messages to the client, soon to come.

We use the new fancy readonly structs with the "in" keyword to effectively create a const ref.

No changes required to the client.